Masuma level 3 completion (#11)
* new sprites

* level 3 + updates

level 3 has been added, level 2 has been modified to have a background and a heart monster to replace the red circle. the monsters now kill the prince upon contact. player has the ability to shoot by pressing the down arrow key and kills the monsters. the game over screen now restarts the level when the player hits restart and starts from the beginning of the game if they press exit. also added the background to level 1.

* lvl 2 update

changed the prince prefab to the one we have

Co-authored-by: Masuma Somji <masuma@nyu.edu>